fix(ci): line-limit maintainability — split oversized adoption module and CLI tests
